requirements.txt
click==6.6
coreapi==2.1.1
decorator==4.0.10
defusedxml==0.4.1
dj-database-url==0.4.1
Django==1.10.4
django-allauth==0.29.0
django-extensions==1.7.5
django-filters==0.2.1
django-leaflet==0.19.0
django-model-utils==2.6
django-rest-auth==0.8.2
django-rest-swagger==2.1.0
djangorestframework==3.5.3
djangorestframework-gis==0.11
djangorestframework-jwt==1.9.0
geocoder==1.19.0
gunicorn==19.6.0
itypes==1.1.0
Markdown==2.6.7
oauthlib==2.0.1
openapi-codec==1.2.1
Pillow==3.4.2
psycopg2==2.6.2
PyJWT==1.4.2
python3-openid==3.0.10
ratelim==0.1.6
requests==2.12.4
requests-oauthlib==0.7.0
simplejson==3.10.0
six==1.10.0
uritemplate==3.0.0
Werkzeug==0.11.11
whitenoise==3.2.2
psycopg2==2.6.2
drf-autodocs==0.4.1
drfdocs==0.0.11
geocoder==1.19.0

